+++ src/main.rs
@@ -6,7 +6,6 @@ use rayon::prelude::*;
 use std::path::Path;
 use std::str::FromStr;
 use std::{fs, path::PathBuf};
-use walkdir::WalkDir;
 
 #[derive(Clone, Copy)]
 pub(crate) enum UseCargoMetadata {
@@ -62,6 +61,10 @@ struct MacheteArgs {
     #[argh(switch)]
     fix: bool,
 
+    /// respect ignore files (.gitignore, .ignore, etc.) when searching for files.
+    #[argh(switch)]
+    ignore: bool,
+
     /// print version.
     #[argh(switch)]
     version: bool,
@@ -71,21 +74,26 @@ struct MacheteArgs {
     paths: Vec<PathBuf>,
 }
 
-fn collect_paths(path: &Path, skip_target_dir: bool) -> Result<Vec<PathBuf>, walkdir::Error> {
-    // Find directory entries.
-    let walker = WalkDir::new(path).into_iter();
+struct CollectPathOptions {
+    skip_target_dir: bool,
+    respect_ignore_files: bool,
+}
 
-    let manifest_path_entries = if skip_target_dir {
-        walker
-            .filter_entry(|entry| !entry.path().ends_with("target"))
-            .collect()
-    } else {
-        walker.collect::<Vec<_>>()
-    };
+fn collect_paths(path: &Path, options: CollectPathOptions) -> Result<Vec<PathBuf>, ignore::Error> {
+    // Find directory entries.
+    let mut builder = ignore::WalkBuilder::new(path);
+
+    builder.standard_filters(options.respect_ignore_files);
+
+    if options.skip_target_dir {
+        builder.filter_entry(|entry| !entry.path().ends_with("target"));
+    }
+
+    let walker = builder.build();
 
     // Keep only errors and `Cargo.toml` files (filter), then map correct paths into owned
     // `PathBuf`.
-    manifest_path_entries
+    walker
         .into_iter()
         .filter(|entry| match entry {
             Ok(entry) => entry.file_name() == "Cargo.toml",
@@ -142,7 +150,13 @@ fn run_machete() -> anyhow::Result<bool> {
     let mut walkdir_errors = Vec::new();
 
     for path in args.paths {
-        let manifest_path_entries = match collect_paths(&path, args.skip_target_dir) {
+        let manifest_path_entries = match collect_paths(
+            &path,
+            CollectPathOptions {
+                skip_target_dir: args.skip_target_dir,
+                respect_ignore_files: args.ignore,
+            },
+        ) {
             Ok(entries) => entries,
             Err(err) => {
                 walkdir_errors.push(err);
@@ -251,7 +265,7 @@ fn run_machete() -> anyhow::Result<bool> {
 }
 
 fn remove_dependencies(manifest: &str, dependencies_list: &[String]) -> anyhow::Result<String> {
-    let mut manifest = toml_edit::Document::from_str(manifest)?;
+    let mut manifest = toml_edit::DocumentMut::from_str(manifest)?;
     let dependencies = manifest
         .iter_mut()
         .find_map(|(k, v)| (v.is_table_like() && k == "dependencies").then_some(Some(v)))
@@ -313,13 +327,28 @@ const TOP_LEVEL: &str = concat!(env!("CARGO_MANIFEST_DIR"));
 fn test_ignore_target() {
     let entries = collect_paths(
         &PathBuf::from(TOP_LEVEL).join("./integration-tests/with-target/"),
-        true,
+        CollectPathOptions {
+            skip_target_dir: true,
+            respect_ignore_files: false,
+        },
     );
     assert!(entries.unwrap().is_empty());
 
     let entries = collect_paths(
         &PathBuf::from(TOP_LEVEL).join("./integration-tests/with-target/"),
-        false,
+        CollectPathOptions {
+            skip_target_dir: false,
+            respect_ignore_files: true,
+        },
+    );
+    assert!(entries.unwrap().is_empty());
+
+    let entries = collect_paths(
+        &PathBuf::from(TOP_LEVEL).join("./integration-tests/with-target/"),
+        CollectPathOptions {
+            skip_target_dir: false,
+            respect_ignore_files: false,
+        },
     );
     assert!(!entries.unwrap().is_empty());
 }